requirements.txt
boto3==1.7.41
requests==2.19.1
websocket-client==0.48.0
py_zipkin==0.12.0
pycrypto==2.6.1